% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{get_bbox}
\alias{get_bbox}
\title{Extract IMD data within a bounding box}
\usage{
get_bbox(
  lat_min,
  lat_max,
  lon_min,
  lon_max,
  variable,
  start_yr,
  end_yr,
  file_dir = ".",
  format = "netcdf",
  save = TRUE
)
}
\arguments{
\item{lat_min}{Minimum latitude.}

\item{lat_max}{Maximum latitude.}

\item{lon_min}{Minimum longitude.}

\item{lon_max}{Maximum longitude.}

\item{variable}{One of "rain", "tmax", "tmin".}

\item{start_yr}{Start year.}

\item{end_yr}{End year.}

\item{file_dir}{Directory for files.}

\item{format}{"netcdf" (default) or "geotiff".}

\item{save}{Save output? Default TRUE.}
}
\value{
Invisible SpatRaster of the cropped region.
}
\description{
Extract IMD data within a bounding box
}
