% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{get_point_all}
\alias{get_point_all}
\title{Extract daily time series for all variables at a point}
\usage{
get_point_all(lat, lon, start_yr, end_yr, file_dir = ".", save_csv = TRUE)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}

\item{start_yr}{Start year.}

\item{end_yr}{End year.}

\item{file_dir}{Directory for .grd files.}

\item{save_csv}{Save merged output as CSV? Default TRUE.}
}
\value{
Invisible data frame with columns date, lat, lon,
rain, tmax, tmin, dtr.
}
\description{
Extract daily time series for all variables at a point
}
