% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{open_data}
\alias{open_data}
\title{Read cached IMD .grd files from disk}
\usage{
open_data(variable, start_yr, end_yr, file_dir = ".")
}
\arguments{
\item{variable}{One of "rain", "tmax", "tmin".}

\item{start_yr}{Start year.}

\item{end_yr}{End year.}

\item{file_dir}{Directory containing the variable sub-folder.}
}
\value{
A SpatRaster (single year) or named list (multi-year).
}
\description{
Read cached IMD .grd files from disk
}
\examples{
\dontrun{
rain2020 <- open_data("rain", 2020, 2020, "~/imdR_data")
rain_3yr <- open_data("rain", 2018, 2020, "~/imdR_data")
}
}
