% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary.R
\name{read_imd_binary}
\alias{read_imd_binary}
\title{Read an IMD binary .grd file}
\usage{
read_imd_binary(filepath, variable, year)
}
\arguments{
\item{filepath}{Path to the .grd file.}

\item{variable}{One of "rain", "tmax", "tmin".}

\item{year}{Integer year.}
}
\value{
A numeric 3-dimensional array of dimensions ncols x nrows x ndays.
}
\description{
Read an IMD binary .grd file
}
\keyword{internal}
