% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_geotiff}
\alias{to_geotiff}
\title{Save an IMD SpatRaster as a compressed GeoTIFF}
\usage{
to_geotiff(imd_raster, file_path)
}
\arguments{
\item{imd_raster}{A terra SpatRaster.}

\item{file_path}{Output .tif file path.}
}
\value{
Invisible file path.
}
\description{
Save an IMD SpatRaster as a compressed GeoTIFF
}
