% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_netcdf}
\alias{to_netcdf}
\title{Save an IMD SpatRaster as CF-1.7 compliant NetCDF}
\usage{
to_netcdf(imd_raster, file_path, variable = "rain")
}
\arguments{
\item{imd_raster}{A terra SpatRaster.}

\item{file_path}{Output .nc file path.}

\item{variable}{One of "rain", "tmax", "tmin".}
}
\value{
Invisible file path.
}
\description{
Save an IMD SpatRaster as CF-1.7 compliant NetCDF
}
