% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd_postcode_lookup.R
\name{imd_lookup}
\alias{imd_lookup}
\title{Lookup IMD Using Postcode}
\usage{
imd_lookup(postcodes, pcdcolumn)
}
\arguments{
\item{postcodes}{This is the dataframe containing all data}

\item{pcdcolumn}{This isd the name of the column that contains the postcode data}
}
\description{
Function for looking up a column of postcodes and returning information about the
indices of deprivation about that area
}
