% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_csv}
\alias{to_csv}
\title{Extract a daily time series at a point location}
\usage{
to_csv(imd_raster, lat, lon, file_path = NULL)
}
\arguments{
\item{imd_raster}{A terra SpatRaster from get_data().}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}

\item{file_path}{If provided, saves output as CSV.}
}
\value{
Invisible data frame with columns date and value.
}
\description{
Extract a daily time series at a point location
}
