% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{to_netcdf}
\alias{to_netcdf}
\title{Save an IMD SpatRaster as CF-1.7 compliant NetCDF}
\usage{
to_netcdf(imd_raster, file_path, variable = "rain")
}
\arguments{
\item{imd_raster}{A terra SpatRaster.}

\item{file_path}{Output .nc file path.}

\item{variable}{One of "rain", "tmax", "tmin".}
}
\value{
Invisible file path.
}
\description{
Save an IMD SpatRaster as CF-1.7 compliant NetCDF
}
\examples{
\dontrun{
r <- get_data("rain", 2020, 2020, "~/imdR_data")
to_netcdf(r, "~/imdR_data/rain_2020.nc", "rain")

# Save a boundary-extracted region
goa <- extract_by_boundary(r, "state", "Goa", "rain")
to_netcdf(goa, "~/imdR_data/rain_Goa_2020.nc", "rain")
}
}
